Add top level exports to TypeScript defs
